requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9    # required for FastAPI file uploads

# LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.0

# Vector store
faiss-cpu>=1.7.4

# Embeddings (free, local — no API key required)
sentence-transformers>=2.7.0

# Document loaders
pypdf>=4.0.0
docx2txt>=0.8

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
